requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
flake8==6.1.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
sqlparse==0.4.4
testfixtures==7.2.0
tzdata==2023.3
uritemplate==4.1.1
